refactor: 컴포넌트 분할 및 일부 리팩토링 진행
