+++ fourtoon-cookie/src/pages/SettingPage/Footer/Footer.styled.ts
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+
+export const styles = StyleSheet.create({
+    footer: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        marginTop: 16,
+      },
+      inquiryText: {
+        fontSize: 16,
+        color: '#aaa',
+        textDecorationLine: 'underline',
+        textDecorationColor: '#ddd',
+      },
+      logoutText: {
+        fontSize: 16,
+        color: 'red',
+        textDecorationLine: 'underline',
+        textDecorationColor: 'red',
+      },
+});
