+++ fourtoon-cookie/src/pages/SettingPage/LoginInfoLayout/LoginInfoLayout.styled.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    infoText: {
+        fontSize: 16,
+    }
+});
